feat: :sparkles: handle delete globally
